feat: Bump react-navigation to v5 with code adjustments
