+++ package.json
@@ -56,14 +56,14 @@
   "licenseFilename": "LICENSE",
   "readmeFilename": "README.md",
   "peerDependencies": {
-    "@react-navigation/core": ">=4.0.0",
-    "feature-u": ">=1.0.0",
+    "@react-navigation/core": "^5.15.0",
+    "feature-u": "^3.0.0",
     "react": ">=16.9.0"
   },
   "dependencies": {
     "@commitlint/prompt": "^11.0.0",
     "lodash.has": "^4.5.2",
-    "react-is": "^16.12.0"
+    "react-is": "^17.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.7.7",
@@ -74,7 +74,7 @@
     "@babel/runtime": "^7.7.7",
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
-    "@react-navigation/core": "^3.5.1",
+    "@react-navigation/core": "^5.15.0",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "babel-jest": "^26.5.2",
